Revert "Drools 2851 Verifier: Add support for ranges (#2171)"
This reverts commit 704ba3d728a8df0c8056cb3d8d67e6c5cfdaa746.
