+++ drools-verifier/drools-verifier-api/src/main/resources/META-INF/ErraiApp.properties
@@ -18,8 +18,6 @@ errai.marshalling.serializableTypes=org.drools.verifier.api.reporting.CheckType 
   org.drools.verifier.api.reporting.ImpossibleMatchIssue \
   org.drools.verifier.api.reporting.Issue \
   org.drools.verifier.api.reporting.Issues \
-  org.drools.verifier.api.reporting.gaps.Partition \
-  org.drools.verifier.api.reporting.gaps.MissingRangeIssue \
   org.drools.verifier.api.reporting.MultipleValuesForOneActionIssue \
   org.drools.verifier.api.reporting.RedundantConditionsIssue \
   org.drools.verifier.api.reporting.Severity \
