+++ drools-verifier/drools-verifier-core/src/main/java/org/drools/verifier/core/checks/DetectConflictingRowsCheck.java
@@ -17,9 +17,7 @@
 package org.drools.verifier.core.checks;
 
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.HashSet;
-import java.util.List;
 
 import org.drools.verifier.api.reporting.CheckType;
 import org.drools.verifier.api.reporting.Issue;
@@ -56,9 +54,8 @@ public class DetectConflictingRowsCheck
     }
 
     @Override
-    protected List<Issue> makeIssues(final Severity severity,
-                                     final CheckType checkType) {
-
+    protected Issue makeIssue(final Severity severity,
+                              final CheckType checkType) {
         final Issue issue = new Issue(severity,
                                       checkType,
                                       new HashSet<>(Arrays.asList(ruleInspector.getRowIndex() + 1,
@@ -67,6 +64,6 @@ public class DetectConflictingRowsCheck
 
         issue.setDebugMessage(new RuleInspectorDumper(ruleInspector).dump() + " ## " + new RuleInspectorDumper(other).dump());
 
-        return Collections.singletonList(issue);
+        return issue;
     }
 }
