+++ drools-verifier/drools-verifier-core/src/main/java/org/drools/verifier/core/checks/DetectDeficientRowsCheck.java
@@ -16,9 +16,8 @@
 
 package org.drools.verifier.core.checks;
 
-import java.util.Collections;
+import java.util.Arrays;
 import java.util.HashSet;
-import java.util.List;
 
 import org.drools.verifier.api.reporting.CheckType;
 import org.drools.verifier.api.reporting.Issue;
@@ -62,13 +61,11 @@ public class DetectDeficientRowsCheck
     }
 
     @Override
-    protected List<Issue> makeIssues(final Severity severity,
-                                     final CheckType checkType) {
-        return Collections.singletonList(
-                new Issue(severity,
-                          checkType,
-                          new HashSet<>(Collections.singleton(ruleInspector.getRowIndex() + 1))
-                )
+    protected Issue makeIssue(final Severity severity,
+                              final CheckType checkType) {
+        return new Issue(severity,
+                         checkType,
+                         new HashSet<>(Arrays.asList(ruleInspector.getRowIndex() + 1))
         );
     }
 }
