+++ drools-verifier/drools-verifier-core/src/main/java/org/drools/verifier/core/checks/DetectMissingConditionCheck.java
@@ -18,7 +18,6 @@ package org.drools.verifier.core.checks;
 
 import java.util.Arrays;
 import java.util.HashSet;
-import java.util.List;
 
 import org.drools.verifier.api.reporting.CheckType;
 import org.drools.verifier.api.reporting.Issue;
@@ -48,13 +47,11 @@ public class DetectMissingConditionCheck
     }
 
     @Override
-    protected List<Issue> makeIssues(final Severity severity,
-                                     final CheckType checkType) {
-        return Arrays.asList(
-                new Issue(severity,
-                          checkType,
-                          new HashSet<>(Arrays.asList(ruleInspector.getRowIndex() + 1))
-                )
+    protected Issue makeIssue(final Severity severity,
+                              final CheckType checkType) {
+        return new Issue(severity,
+                         checkType,
+                         new HashSet<>(Arrays.asList(ruleInspector.getRowIndex() + 1))
         );
     }
 }
