+++ drools-verifier/drools-verifier-core/src/main/java/org/drools/verifier/core/checks/DetectMultipleValuesForOneActionCheck.java
@@ -16,10 +16,8 @@
 
 package org.drools.verifier.core.checks;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashSet;
-import java.util.List;
 
 import org.drools.verifier.api.reporting.CheckType;
 import org.drools.verifier.api.reporting.Issue;
@@ -63,16 +61,12 @@ public class DetectMultipleValuesForOneActionCheck
     }
 
     @Override
-    protected List<Issue> makeIssues(final Severity severity,
-                                     final CheckType checkType) {
-        final ArrayList<Issue> result = new ArrayList<Issue>();
-
-        result.add(new MultipleValuesForOneActionIssue(severity,
-                                                       checkType,
-                                                       HumanReadable.toHumanReadableString(conflict.getConflictedItem()),
-                                                       HumanReadable.toHumanReadableString(conflict.getConflictingItem()),
-                                                       new HashSet<>(Arrays.asList(ruleInspector.getRowIndex() + 1))));
-
-        return result;
+    protected Issue makeIssue(final Severity severity,
+                              final CheckType checkType) {
+        return new MultipleValuesForOneActionIssue(severity,
+                                                   checkType,
+                                                   HumanReadable.toHumanReadableString(conflict.getConflictedItem()),
+                                                   HumanReadable.toHumanReadableString(conflict.getConflictingItem()),
+                                                   new HashSet<>(Arrays.asList(ruleInspector.getRowIndex() + 1)));
     }
 }
