+++ drools-verifier/drools-verifier-core/src/main/java/org/drools/verifier/core/checks/base/Check.java
@@ -16,8 +16,6 @@
 
 package org.drools.verifier.core.checks.base;
 
-import java.util.List;
-
 import org.drools.verifier.api.reporting.Issue;
 import org.drools.verifier.core.configuration.CheckConfiguration;
 
@@ -25,7 +23,7 @@ public interface Check {
 
     boolean check();
 
-    List<Issue> getIssues();
+    Issue getIssue();
 
     boolean hasIssues();
