+++ drools-verifier/drools-verifier-core/src/main/java/org/drools/verifier/core/checks/base/PriorityListCheck.java
@@ -42,8 +42,8 @@ public class PriorityListCheck
     }
 
     @Override
-    public List<Issue> getIssues() {
-        return checkWithIssues.getIssues();
+    public Issue getIssue() {
+        return checkWithIssues.getIssue();
     }
 
     @Override
