--- drools-verifier/drools-verifier-core/src/main/java/org/drools/verifier/core/checks/gaps/BidimensionalRange.java
@@ -1,41 +1,0 @@
-/*
- * Copyright 2019 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.drools.verifier.core.checks.gaps;
-
-public class BidimensionalRange<H extends Comparable, V extends Comparable> {
-
-    private final Range<? extends H> horizontal;
-    private final Range<? extends V> vertical;
-
-    public BidimensionalRange(final Range<? extends H> horizontal,
-                              final Range<? extends V> vertical) {
-        this.horizontal = horizontal;
-        this.vertical = vertical;
-    }
-
-    public Range<? extends H> getHorizontal() {
-        return horizontal;
-    }
-
-    public Range<? extends V> getVertical() {
-        return vertical;
-    }
-
-    @Override
-    public String toString() {
-        return "[" + horizontal + "][" + vertical + "]";
-    }
-}
