--- drools-verifier/drools-verifier-core/src/main/java/org/drools/verifier/core/checks/gaps/OperatorType.java
@@ -1,38 +1,0 @@
-/*
- * Copyright 2019 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.drools.verifier.core.checks.gaps;
-
-import org.drools.verifier.core.relations.Operator;
-
-public enum OperatorType {
-    PARTITION,
-    RANGE,
-    UNKNOWN;
-
-    static OperatorType decode(final Operator op) {
-        return op == Operator.EQUALS ? PARTITION : op.isRangeOperator() ? RANGE : UNKNOWN;
-    }
-
-    OperatorType combine(final OperatorType other) {
-        if (this == UNKNOWN || other == UNKNOWN) {
-            return UNKNOWN;
-        }
-        if (this == RANGE || other == RANGE) {
-            return RANGE;
-        }
-        return PARTITION;
-    }
-}
